restructure pages && redesigned lists
