+++ next.config.js
@@ -11,15 +11,6 @@ const nextConfig = {
       },
     ],
   },
-  async redirects() {
-    return [
-      {
-        source: "/",
-        destination: "/dashboard",
-        permanent: false,
-      },
-    ];
-  },
 };
 
 module.exports = nextConfig;
