--- src/app/(dashboard)/dashboard/page.tsx
@@ -1,9 +1,0 @@
-import Dashboard from "@/components/page/dashboard/Dashboard";
-
-export default function page() {
-  return (
-    <>
-      <Dashboard />
-    </>
-  );
-}
