+++ src/app/(main)/details/page.tsx
