+++ src/app/(main)/page.tsx
@@ -1,0 +1,9 @@
+import List from "@/components/page/list/List";
+
+export default function page() {
+  return (
+    <>
+      <List />
+    </>
+  );
+}
