--- src/app/page.tsx
@@ -1,3 +1,0 @@
-export default function page() {
-  return null;
-}
