+++ src/components/page/list/Card.module.css
@@ -1,0 +1,8 @@
+.detail__box {
+  width: 22vw;
+  display: flex;
+  flex-direction: column;
+  border-radius: 1rem;
+  margin: 2rem 0.5rem;
+  padding: 1rem;
+}
