+++ src/components/page/list/CardDetails.module.css
@@ -1,22 +1,10 @@
-.detail__box {
-  width: 22vw;
-  display: flex;
-  flex-direction: column;
-  border-radius: 1rem;
-  margin: 2rem 0.5rem;
-  padding: 1rem;
-}
 
-.detail__box:hover {
-  cursor: pointer;
-}
 .details {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   height: 15vh;
   color: rgba(33, 33, 33);
-  margin: 0 1rem;
 }
 
 .location {
@@ -25,22 +13,24 @@
   margin: 1rem 0;
   overflow: hidden;
 }
-.location h2 {
-  font-size: 2rem;
-  font-weight: 500;
+.location h1 {
+  font-size: 1.5rem;
+  font-weight: bolder;
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
 }
 
-.location h3 {
+.location h2 {
   font-size: 1.5rem;
-  font-weight: 400;
-  color: rgba(33, 33, 33, 0.9);
+  font-weight: bold;
+  color: rgba(33, 33, 33, 0.7);
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
+  margin: .5rem 0;
 }
+
 .other__details {
   display: flex;
   justify-content: space-between;
@@ -51,7 +41,7 @@
   color: rgba(33, 33, 33);
 
   font-size: 1.4rem;
-  font-weight: 500;
+  font-weight: bold;
 }
 
 .price {
@@ -59,7 +49,8 @@
   align-items: center;
 }
 .unknown {
-  font-size: 1.2rem;
+  font-size: 1.5rem;
+  font-weight: bold;
   font-style: italic;
   color: rgba(33, 33, 33, 0.6);
   margin: 0 1rem;
