+++ src/components/page/list/CardImage.module.css
@@ -1,0 +1,40 @@
+.img__outer__container {
+  width: 100%;
+  aspect-ratio: calc(1);
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 0.5rem;
+}
+
+.img__container {
+  border-radius: 1rem;
+
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+}
+
+.img {
+  object-fit: cover;
+  width: 20vw;
+  height: auto;
+  aspect-ratio: calc(1);
+  border-radius: 0.5rem;
+}
+
+
+.no__img {
+  font-size: 5rem;
+  color: #222222;
+}
+
+.location__icon {
+  position: absolute;
+  z-index: 20;
+  left: 50%;
+  top: 50%;
+}
